feat: Setup a proxmox cluster on scaleway servers (#538)
This PR adds a proxmox cluster on scaleway servers.
Those servers are not yet used in production.

# Remarks

- On each machine, inside `proxmox_node__pve_storages`, I had to remove
`shared: 0` on the zfs-hdd-pve` as it made the ansible setup crash. I'm
not sure why, maybe related to this table where shared is marked as
incompatible with ZFS?
https://pve.proxmox.com/wiki/Storage#_storage_types
- AFAIK `proxmox_node__pve_cluster_name` must equal the Ansible
inventory group name. So we can't have a `-` in it. It is also a proxmox
cluster name, so it must not have a `_` in it. Therefor, I choose
`pvescaleway` for `proxmox_node__pve_cluster_name`. What surprises me is
that `pve-hetzner` worked for hetzner? It doesn't equal the inventory
group name, which is weird.